+++ XXXX-dao/src/main/java/com/framework/dao/mapper/system/SystemBlackListIpMapper.java
@@ -42,7 +42,7 @@ public interface SystemBlackListIpMapper {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insert(SystemBlackListIp record);
+    int insert(SystemBlackListIp record);
 
     /**
      * @param record 1 系统前端操作黑名单IP实体类对象
@@ -52,7 +52,7 @@ public interface SystemBlackListIpMapper {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insertSelective(SystemBlackListIp record);
+    int insertSelective(SystemBlackListIp record);
 
 
     /**
