+++ XXXX-dao/src/main/java/com/framework/dao/mapper/system/SystemButtonMapper.java
@@ -42,7 +42,7 @@ public interface SystemButtonMapper {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insert(SystemButton record);
+    int insert(SystemButton record);
 
     /**
      * @param record 1 按钮实体类对象
@@ -52,7 +52,7 @@ public interface SystemButtonMapper {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insertSelective(SystemButton record);
+    int insertSelective(SystemButton record);
 
     /**
      * @param list 1 按钮批量添加集合
