+++ XXXX-service/src/main/java/com/framework/service/service/other/FileInfoService.java
@@ -42,7 +42,7 @@ public interface FileInfoService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insert(FileInfo record);
+    int insert(FileInfo record);
 
     /**
      * @param record 1 文件实体类对象
@@ -52,7 +52,7 @@ public interface FileInfoService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insertSelective(FileInfo record);
+    int insertSelective(FileInfo record);
 
     /**
      * @param list 1 文件批量添加集合
