+++ XXXX-service/src/main/java/com/framework/service/service/system/SystemRoleService.java
@@ -32,7 +32,7 @@ public interface SystemRoleService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insert(SystemRole record);
+    int insert(SystemRole record);
 
     /**
      * @param record 1 角色实体类对象
@@ -42,7 +42,7 @@ public interface SystemRoleService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insertSelective(SystemRole record);
+    int insertSelective(SystemRole record);
 
     /**
      * @param list 1 角色批量添加集合
