+++ XXXX-service/src/main/java/com/framework/service/service/system/SystemTypeService.java
@@ -32,7 +32,7 @@ public interface SystemTypeService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insert(SystemType record);
+    int insert(SystemType record);
 
     /**
      * @param record 1 类型实体类对象
@@ -42,7 +42,7 @@ public interface SystemTypeService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insertSelective(SystemType record);
+    int insertSelective(SystemType record);
 
 
     /**
