+++ XXXX-service/src/main/java/com/framework/service/service/system/SystemUserService.java
@@ -32,7 +32,7 @@ public interface SystemUserService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insert(SystemUser record);
+    int insert(SystemUser record);
 
     /**
      * @param record 1 用户实体类对象
@@ -42,7 +42,7 @@ public interface SystemUserService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insertSelective(SystemUser record);
+    int insertSelective(SystemUser record);
 
     /**
      * @param list 1 用户批量添加集合
