+++ XXXX-service/src/main/java/com/framework/service/service/system/SystemWhiteListIpService.java
@@ -32,7 +32,7 @@ public interface SystemWhiteListIpService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insert(SystemWhiteListIp record);
+    int insert(SystemWhiteListIp record);
 
     /**
      * @param record 1 系统后台操作白名单IP实体类对象
@@ -42,7 +42,7 @@ public interface SystemWhiteListIpService {
      * @Author 邋遢龘鵺
      * @DateTime 2019/12/26 9:31
      */
-    long insertSelective(SystemWhiteListIp record);
+    int insertSelective(SystemWhiteListIp record);
 
 
     /**
